+++ lib/config.php
@@ -9,7 +9,7 @@ return array(
 	/*
 	 * Application version
 	 */
-	'version' => '0.7.2',
+	'version' => '0.7.3',
 
 	/*
 	 * PHP minimum version