+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to Pug will be documented in this file (beginning with v0.5 😅).
 
+## [Unreleased]
+### Added
+- Specify alternative storage location using `PUGFILE` environment variable
+
 ## [0.7.2] - 2016-12-23
 ### Added
 - Support for displaying project metadata